Keep whitespace lines again, fix getIndent()
I came to the realization that while removing whitespace lines makes
token navigation 'less' complicated, it removed the ability to determine
line numbers on any pass but the first (a.k.a. will break when
retract()'ing). So keep that stuff.

getIndent() was experiencing a bug with incomplete tabs where it would
pass LineToken tests, but fail in the Parser, which hasn't been tested
yet. This replaces the implementation of getIndent with a much more
succinct version that actually works.