Fix _parse() calling Lexer.next() while unwinding
TL;DR Parse._parse() was fixed, Lexer.retract() returns new current
token.

While adding an error for calling `Lexer.next()` when the current token
is the ending EOFTOKEN, it became apparent that there were a number of
flaws with the flow of the parse subroutine. Notably:

- When the EOFTOKEN was encountered while in a nested block, each return
  from the recursive `Parser._parse()` call would then call `next()` on
  the Lexer, advancing the pointer for each level of recursion.

- Even changing the order wasn't enough to fix it, because while
  unraveling the recursion the value of `token` was not preserved. So,
  the value of `token` upon return would be whatever the token was before
  the recursive call to _parse(), resulting in not detecting EOF
  properly. This has been fixed by always using the internal
  lexer's current token state, which is persistent throughout
  the recursion.

- Also, when calling `Lexer.retract()`, the token value was not being
  updated appropriately. Honestly, it's a miracle it worked at all. This
  function has been updated to return the value of the ending `next()`
  call, so as to parallel the behavior of `next()`.

Even once the actual Lexer worked correctly, some of the test cases were
not valid inputs in significant ways. For example, the previous test for
"retract() out of range" use an incomplete class definition without a
body, causing the test case to not behave as expected, this has also
been fixed.