Unravel enough times when indent difference is big
When the difference between the indent at the leaves of a block and the
indent of the next block differed by more than 1, it would not unravel
properly. It should traverse up the tree by the difference between the
leaf indent and the new block indent in order to place the next block
properly within the tree. In order to facilitate this, a global indent
level state must be used, so that the indent can be modified before
entering the next _parse invocation.